fix sh

+++ install.sh
@@ -3,7 +3,6 @@ for f in .??*
   do
     [[ "$f" == ".git" ]] && continue
     [[ "$f" == ".DS_Store" ]] && continue
-  
-    mv ~/$f ~/dotfiles/oldfiles/
+    [ -f ~/$f ] && mv ~/$f ~/dotfiles/oldfiles/
     ln -s ~/dotfiles/$f ~/$f 
   done
